refactor deepseek evals
