+++ examples/rl_gsm8k/deepseek_math_eval/README.md
@@ -1,0 +1,3 @@
+# Deepseek Math Eval
+
+The evaluation scripts are taken from the [Deepseek-Math](https://github.com/deepseek-ai/DeepSeek-Math/) repository. The scripts are modified to work with the TapeAgents framework.
